Rama lea

+++ src/main/java/com/gimnasio/demo/Service/UsuarioServicio.java
@@ -2,6 +2,7 @@ package com.gimnasio.demo.Service;
 
 import com.gimnasio.demo.Controller.UsuarioController;
 import com.gimnasio.demo.DTO.UsuarioRegistroDTO;
+import com.gimnasio.demo.Exceptions.UsuarioNoEncontradoException;
 import com.gimnasio.demo.Model.Tarjeta;
 import com.gimnasio.demo.Repository.TarjetaRepositorio;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,11 +25,18 @@ public class UsuarioServicio {
     @Autowired
     private TarjetaRepositorio tarjetaRepositorio;
 
-    public Optional<Usuario> buscarUsuarioPorID(Long id){
-        return usuarioRepositorio.findById(id);
+    public Optional<Usuario> buscarUsuarioPorID(Long id) throws UsuarioNoEncontradoException{
+        Optional<Usuario> usuario;
+        if(usuarioRepositorio.existsById(id)){
+             usuario=usuarioRepositorio.findById(id);
+        }else{
+            throw new UsuarioNoEncontradoException("ese usuario no existe");
+        }
+
+        return usuario;
     }
 
-    public Usuario convertidorDTO(UsuarioRegistroDTO usu){
+    public Usuario conversorDTO(UsuarioRegistroDTO usu){
         Usuario usuu = new Usuario(usu.getUsername(), usu.getEmail(), usu.getContrasena(), usu.getApellido(), usu.getNombre(), usu.getDni(), usu.getDomicilio());
         return usuu;
     }
@@ -36,7 +44,7 @@ public class UsuarioServicio {
     public boolean crearUsuario(UsuarioRegistroDTO Dto)
     {
         boolean b=false;
-        Usuario usu= convertidorDTO(Dto);
+        Usuario usu= conversorDTO(Dto);
 
         if(!usuarioRepositorio.existsByEmail(usu.getEmail()) && !usuarioRepositorio.existsByDni(usu.getDni()) && !usuarioRepositorio.existsByUsername(usu.getUsername())){
             b=true;
@@ -46,20 +54,24 @@ public class UsuarioServicio {
         return b;
     }
 
-    public void deleteUser(Usuario usuario){usuarioRepositorio.delete(usuario);}
+    public void eliminarUsuarioPorID(Long id) throws UsuarioNoEncontradoException{
 
-    /// IMPLEMENTAR EXCEPTION/////////////////////////////////////////////////////////////////////////////////////
-    public void eliminarUsuarioPorID(Long id){
-        usuarioRepositorio.deleteById(id);
+        if(usuarioRepositorio.existsById(id)){
+            usuarioRepositorio.deleteById(id);
+        }else{
+            throw new UsuarioNoEncontradoException("ese usuario no existe");
+        }
     }
 
     public List<Usuario> listarUsuarios(){
         return usuarioRepositorio.findAll();
     }
 
-    public void editarUsuario(Long id, Usuario usuario){
+    public void editarUsuario(Long id, Usuario usuario) throws UsuarioNoEncontradoException{
         if(usuarioRepositorio.existsById(id)){
             usuarioRepositorio.save(usuario);
+        }else{
+            throw new UsuarioNoEncontradoException("ese usuario no existe");
         }
     }
 
@@ -72,7 +84,4 @@ public class UsuarioServicio {
 
         return Optional.of(tarjetas);
     }
-
-
-    
 }
